add first attempt at LZSH

+++ src/determinant.cpp
@@ -47,12 +47,12 @@ std::vector<Determinant> Determinant::excitations(const std::vector<int>& excs) 
 
     // single excitations
     if (std::find(excs.begin(), excs.end(), 1) != excs.end()) {
-        for (int i = 0; i < a.size(); i++) {
+        for (size_t i = 0; i < a.size(); i++) {
             for (int j = a.size(); j < norb; j++) {
                 std::vector<int> det = this->a; det.at(i) = j; excitations.push_back(Determinant(norb, det, b));
             }
         }
-        for (int i = 0; i < b.size(); i++) {
+        for (size_t i = 0; i < b.size(); i++) {
             for (int j = b.size(); j < norb; j++) {
                 std::vector<int> det = this->b; det.at(i) = j; excitations.push_back(Determinant(norb, a, det));
             }
@@ -61,27 +61,27 @@ std::vector<Determinant> Determinant::excitations(const std::vector<int>& excs) 
 
     // double excitations
     if (std::find(excs.begin(), excs.end(), 2) != excs.end()) {
-        for (int i = 0; i < a.size(); i++) {
+        for (size_t i = 0; i < a.size(); i++) {
             for (int j = a.size(); j < norb; j++) {
-                for (int k = 0; k < b.size(); k++) {
+                for (size_t k = 0; k < b.size(); k++) {
                     for (int l = b.size(); l < norb; l++) {
                         std::vector<int> deta = this->a, detb = this->b; deta.at(i) = j, detb.at(k) = l; excitations.push_back(Determinant(norb, deta, detb));
                     }
                 }
             }
         }
-        for (int i = 0; i < a.size(); i++) {
+        for (size_t i = 0; i < a.size(); i++) {
             for (int j = a.size(); j < norb; j++) {
-                for (int k = i + 1; k < a.size(); k++) {
+                for (size_t k = i + 1; k < a.size(); k++) {
                     for (int l = j + 1; l < norb; l++) {
                         std::vector<int> det = this->a; det.at(i) = j, det.at(k) = l; excitations.push_back(Determinant(norb, det, b));
                     }
                 }
             }
         }
-        for (int i = 0; i < b.size(); i++) {
+        for (size_t i = 0; i < b.size(); i++) {
             for (int j = b.size(); j < norb; j++) {
-                for (int k = i + 1; k < b.size(); k++) {
+                for (size_t k = i + 1; k < b.size(); k++) {
                     for (int l = j + 1; l < norb; l++) {
                         std::vector<int> det = this->b; det.at(i) = j, det.at(k) = l; excitations.push_back(Determinant(norb, a, det));
                     }
